# Improve `__clzsi2` performance