This was something I originally made for PR 332, but I realized I should PR it by itself so I can remove it from my `specialized-div-rem` crate. As an example, the assembly generation for RISC-V used to be
```
 addi    a1, zero, 64
 beqz    a2, .LBB202_2
 addi    a1, zero, 32
 add     a0, zero, a2
.LBB202_2:
 srli    a2, a0, 16
 beqz    a2, .LBB202_7
 addi    a1, a1, -16
 srli    a0, a2, 8
 bnez    a0, .LBB202_8
.LBB202_4:
 add     a0, zero, a2
 srli    a2, a0, 4
 bnez    a2, .LBB202_9
.LBB202_5:
 add     a2, zero, a0
 srli    a0, a2, 2
 bnez    a0, .LBB202_10
.LBB202_6:
 add     a0, zero, a2
 addi    a3, zero, 1
 addi    a2, zero, -2
 bgeu    a3, a0, .LBB202_11
 j       .LBB202_12
.LBB202_7:
 add     a2, zero, a0
 srli    a0, a2, 8
 beqz    a0, .LBB202_4
.LBB202_8:
 addi    a1, a1, -8
 srli    a2, a0, 4
 beqz    a2, .LBB202_5
.LBB202_9:
 addi    a1, a1, -4
 srli    a0, a2, 2
 beqz    a0, .LBB202_6
.LBB202_10:
 addi    a1, a1, -2
 addi    a3, zero, 1
 addi    a2, zero, -2
 bltu    a3, a0, .LBB202_12
.LBB202_11:
 neg     a2, a0
.LBB202_12:
 add     a0, a2, a1
 ret
```
but is now:
```
 srli    a1, a0, 32
 snez    a1, a1
 slli    a1, a1, 5
 srl     a0, a0, a1
 srli    a2, a0, 16
 snez    a2, a2
 slli    a2, a2, 4
 srl     a0, a0, a2
 or      a1, a1, a2
 addi    a2, zero, 255
 sltu    a2, a2, a0
 slli    a2, a2, 3
 srl     a0, a0, a2
 or      a1, a1, a2
 addi    a2, zero, 15
 sltu    a2, a2, a0
 slli    a2, a2, 2
 srl     a0, a0, a2
 or      a1, a1, a2
 addi    a2, zero, 3
 sltu    a2, a2, a0
 slli    a2, a2, 1
 srl     a0, a0, a2
 or      a1, a1, a2
 addi    a2, zero, 1
 sltu    a2, a2, a0
 srl     a0, a0, a2
 or      a1, a1, a2
 add     a0, a0, a1
 addi    a1, zero, 64
 sub     a0, a1, a0
 ret
```
This is completely branchless. In fact, I could make it `const`. A previous comment mentioned making `__clzsi2` `const`, but `intrinsics!` doesn't seem to support `const`.