@alexcrichton 2020-07-13
Thanks! Would you be up to adding a microbenchmark and also expanding the test suite for this intrinsic at the same time as well?