@bjorn3 2020-07-14
You could add benches to https://github.com/rust-lang/compiler-builtins/blob/master/testcrate/tests/count_leading_zeros.rs