@alexcrichton 2020-07-14
There's also examples of benchmarks on https://github.com/rust-lang/compiler-builtins/pull/365, but the general idea is that if this is done for performance reasons it would be good to confirm that the performance did indeed improve.