@AaronKutch 2020-07-14
I'm confused. When I look at [this commit](https://github.com/rust-lang/compiler-builtins/pull/365/commits/0d90538348b5c97f012888340b8eecc0a562edd4), I can see benches being added under a `testcrate/benches` folder. This folder doesn't look like it exists in the github browser or my local clone. When I search for the `#[bench]` keyword, I don't see benchmarks anywhere in `compiler-builtins`. Is there some kind of git submodule magic happening here?