@AaronKutch 2020-07-17
I think that the code is large enough to move the `__clzsi2` function under a `leading_zeros.rs` file instead of being an exceptional case in `mod.rs`, but it seems that the test crate cannot see the new file despite the docs clearly showing one. I presume that `compiler-builtins` is doing something hacky with its exports. I guess we do not want a breaking change to solve such a small inelegance anyway.

My current plan I have of testing both variations of counting leading zeros thoroughly is to have two functions, `usize_leading_zeros_default` and `usize_leading_zeros_riscv`, make `__clzsi2` conditionally use one or the other, and add a feature flag the testcrate can set in order to see both underlying functions.

Would a better solution be having a CI test for RISC-V? I read [this post](https://www.codethink.co.uk/articles/2020/fixing-rusts-test-suite-on-risc-v/) that says "the full Rust test suite passes for a RISC-V cross-compilation toolchain". I see a wide variety of CI tests being performed currently, I wonder why RISC-V isn't included.