@alexcrichton 2020-07-17
I'm not sure what's up with modules, nothing's really that fancy in this crate and it should in general act similarly to the rest of Rust's module system. Are you sure that you reexported the intrinsics in `mod.rs`?

For testing adding RISC-V should be fine.

> I wonder why RISC-V isn't included.

This crate is old. RISC-V targets didn't exist when this crate was created. No one's added it since then.