@AaronKutch 2020-07-23
What do you think about my testing setup? I don't know if there is a better solution. Does `expose-alt-impls` look like a flag that could be used similarly by other specialization tests?