@AaronKutch 2020-07-23
I found out that `u128::leading_zeros` has a huge code size. This PR will help, but I think there is a problem upstream wherever it is defined.
the assembly of
```
pub fn lz(x: u128) -> usize {
    let mut x = x;
    let mut z = 0;
    if ((x >> 64) as u64) != 0 {
        z += 64;
    } else {
        x >>= 64;
    }
    z + x.leading_zeros() as usize
}
```
is huge (much larger than it should be even before this PR) but changing the last line to be
```
   z + (x as u64).leading_zeros() as usize
```
fixes it. I think there is a similar problem somewhere.